refactor(test): centralize CLI harness into tests/common/cli_harness.rs
Move webfang_path(), cmd(), BehavioralTest and the redaction helpers out of
tests/behavioral/main.rs into a shared tests/common/cli_harness.rs so the
behavioral, cli_binary and cli_behavioral test binaries can share one copy.

The assert_snapshot_redacted / assert_snapshot_plain wrappers intentionally stay
at the crate root: insta derives a snapshot's on-disk location from the module
path where assert_snapshot! expands, so relocating them would displace the
existing tests/behavioral/snapshots/ files.
